Create utils package
