+++ src/main/java/exception/InvalidCommandException.java
@@ -1,0 +1,7 @@
+package exception;
+
+public class InvalidCommandException extends DukeException {
+    public InvalidCommandException() {
+        super("Invalid command");
+    }
+}
