+++ src/main/java/exception/InvalidDeadlineException.java
@@ -1,0 +1,7 @@
+package exception;
+
+public class InvalidDeadlineException extends DukeException {
+    public InvalidDeadlineException() {
+        super("Input in the form:\ndeadline {taskName} /by {dueDate}");
+    }
+}
