+++ src/main/java/exception/InvalidIndexException.java
@@ -1,0 +1,7 @@
+package exception;
+
+public class InvalidIndexException extends DukeException {
+    public InvalidIndexException() {
+        super("Index is out of bounds");
+    }
+}
