+++ src/main/java/utils/InputUtil.java
@@ -1,0 +1,17 @@
+package utils;
+
+public class InputUtil {
+    public static int parseIndex(String input) {
+        String[] parts = input.split("\\s+");
+
+        if (parts.length >= 2) {
+            return Integer.parseInt(parts[1]) - 1;
+        }
+
+        return -1;
+    }
+
+    public static String getCommandType(String input) {
+        return input.split(" ", 2)[0];
+    }
+}
