+++ src/main/java/view/Exit.java
@@ -1,7 +1,7 @@
 package view;
 
 import static objects.FilePath.EXIT_PATH;
-import static objects.Utils.getFile;
+import static utils.FileUtil.getFile;
 
 public class Exit {
     public static void display() {
