+++ src/main/java/view/Greeting.java
@@ -2,7 +2,7 @@ package view;
 
 import static objects.FilePath.GREETING_PATH;
 import static objects.FilePath.LOGO_PATH;
-import static objects.Utils.getFile;
+import static utils.FileUtil.getFile;
 
 public class Greeting extends UI {
     public static void display() {
